Add flashplugin detection
Now a warning message label will appear in the service selection dialog that warn you if the flashplugin could not be found (for a service that require flash). If you still want to use that service, a message box will appear that tell you that you have to install the latest flash plugin and you'll be back to the home page.