Fix a bug when changing from one service to another
Sometimes, the current service does not quit when going to another service. I cannot really do anything about that but the user should be able to retry loading the new service's page by going in the service dialog and pressing the new service again. Before this fix, the application would not call load(url) again, the user had to quit the app and try again. Now quitting is not required anymore.