Fix key repeat issue with KGlobalAccell (especially visible with Rdio)
If you keep a media player key pressed, the associated action will be triggered continually. This did not cause a problem with other services since they take some time before completing the play/pause action and cancel any new request, but rdio is reacts quickly and even a simple key press will pause then play the song. The solution is to add a small tempo that prevent new actions to be performed within a 1s delay.