Improve MPRIS control
- Fix MprisPlayer.Position type (qlonglong instead of int)
- Make sure Seeked signal is correctly emitted

Quoting the MPRIS 2 interface documentation:

This signal does not need to be emitted when playback starts or when the track changes, unless the track is starting at an unexpected position. An expected position would be the last known one when going from Paused to Playing, and 0 when going from Stopped to Playing.