Travis: make the distinction between a commit build and a full build
A commit build should be fast: we should just compile, runtests and measure coverage.

A full build occurs when either on a tag or an a commit on the release branch. It compiles the code in release mode, run all tests and create a binary package. The binary package will be uploaded to transfer.sh for every full build and on a github release if on a tag.